break up es compiler a bit, get this working again
