+++ app/searchers/query_grammar/ast/negator.rb
@@ -7,12 +7,6 @@ module QueryGrammar
         @items = items
       end
 
-      def accept visitor
-        items.accept visitor
-
-        visitor.visit_negator self
-      end
-
       def to_h
         {
           not: items.to_h
