+++ app/searchers/query_grammar/compiler/es.rb
@@ -1,141 +1,16 @@
 module QueryGrammar
   class Compiler
-    class Index
-      class IndexDSL
-        class OperatorDSL
-          attr_reader :operator
-
-          def self.build **opts, &block
-            new(**opts).tap do |obj|
-              obj.instance_eval(&block)
-            end.operator
-          end
-
-          def initialize prefix, **opts
-            @operator = {
-              name: nil,
-              description: nil,
-              arity: nil,
-              type: nil,
-              compile: nil
-            }.merge(opts)
-          end
-
-          def respond_to_missing? func, **args
-            return true
-          end
-
-          def method_missing func, *args, **opts, &block
-            @operator[ func ] = args.first || block
-          end
-        end
-
-        def self.build &block
-          new.tap do |obj|
-            obj.instance_eval(&block)
-          end.index
-        end
-
-        def index
-          @index ||= Index.new
-        end
-
-        def type key, &block
-          index.types[ key ] = block
-        end
-
-        def default *args
-          args.each do |field|
-            index.fields[ field.to_s ][:default] = true
-          end
-        end
-
-        def operator prefix, **opts, &block
-          index.operators[ prefix ] = OperatorDSL.build(**opts, &block)
-        end
-
-        def respond_to_missing? func, *args
-          return true if index.types[ func ]
-
-          super(func, *args)
-        end
-
-        def method_missing func, *args, **opts, &block
-          if index.types[ func ]
-            field = {
-              aliases: (opts.delete(:aliases) || []).map(&:to_s),
-              existable: false,
-              sortable: false,
-              default: false,
-              type: func,
-              compile: index.types[ func ]
-            }.merge opts
-
-            index.fields[ args.first.to_s ] = field
-
-            return field
-          end
-
-          super(func, *args, **opts, &block)
-        end
-      end
-
-      def self.build &block
-        IndexDSL.build(&block)
-      end
-
-      def types
-        @types ||= {}
-      end
-
-      def fields
-        @fields ||= {}
-      end
-
-      def operators
-        @operators ||= {}
-      end
-
-      def sortable_fields
-        fields.map do |(field, data)|
-          [ field, data[:aliases] ] if data[:sortable]
-        end.flatten.compact
-      end
-
-      def existable_fields
-        fields.map do |(field, data)|
-          [ field, data[:aliases] ] if data[:existable]
-        end.flatten.compact
-      end
-
-      def default_fields
-        fields.map do |(field, data)|
-          field if data[:default]
-        end.flatten.compact
-      end
-
-      def aliases
-        # alias => field
-        fields.each_with_object({}) do |(field, data), memo|
-          data[:aliases].each do |aliass|
-            memo[ aliass ] = field
-          end
-        end
-      end
-
-      def resolve_field field_or_alias
-        return field_or_alias unless aliases.key? field_or_alias
-        aliases[ field_or_alias ]
-      end
-    end
-
     class ES < Compiler
       class << self
         attr_reader :index
 
         def define_index &block
-          @index = Index.build(&block)
+          @index = QueryGrammar::Index.build(&block)
         end
+      end
+
+      def index
+        self.class.index
       end
 
       def context
@@ -152,8 +27,14 @@ module QueryGrammar
         context
       end
 
-      def index
-        self.class.index
+      protected
+
+      def cloak_compile clause, field=nil
+        field = index.resolve_field(field || clause.prefix)
+        field_data = index.fields[ field ]
+        fail "unable to search on non-existant field #{ field }" unless field_data
+
+        QueryGrammar::Cloaker.new(bind: self).cloak(clause.unary, field, clause.values, &field_data[:compile])
       end
 
       visit :group do |group|
@@ -170,11 +51,7 @@ module QueryGrammar
       end
 
       visit :negator do |negator|
-        if negator.items.is_a?(Array)
-          inside = negator.items.map { |i| i.accept self }.compact
-        else
-          inside = visit negator.items
-        end
+        inside = Array(negator.items).map { |i| i.accept self }.compact
 
         {
           bool: {
@@ -188,8 +65,7 @@ module QueryGrammar
           next {
             bool: {
               should: index.default_fields.map do |field|
-                field_data = index.fields[ index.resolve_field field ]
-                QueryGrammar::Cloaker.new(bind: self).cloak(clause, &field_data[:compile])
+                cloak_compile clause, field
               end
             }
           }
@@ -200,10 +76,7 @@ module QueryGrammar
           next QueryGrammar::Cloaker.new(bind: self).cloak(clause, &handler[:compile])
         end
 
-        field_data = index.fields[ index.resolve_field clause.prefix ]
-        throw "unable to search on non-existant field #{ field }" unless field_data
-
-        QueryGrammar::Cloaker.new(bind: self).cloak(clause, &field_data[:compile])
+        cloak_compile clause
       end
     end
   end
